Add some comments about how the code is invoked
We need to call run_module() twice, once to initialise the COBOL data so we can then pass in the parameters, and then to actually invoke the behaviour. This is a bit messy.

As we refactor, we may get to a point where the COBOL variables are pushed to the edge of the system and this can be simplified in an adaptor.